KeeperNodesStorage
