+++ programs/keeper-utils/KeeperUtils.cpp
@@ -8,6 +8,8 @@
 #include <Coordination/KeeperStateMachine.h>
 #include <Coordination/KeeperStorage.h>
 #include <Coordination/KeeperStorageImpl.h>
+#include <Coordination/KeeperMemNodesStorage.h>
+#include <Coordination/KeeperStorage_fwd.h>
 #include <Common/assert_cast.h>
 #include <Core/ColumnWithTypeAndName.h>
 #include <Core/ColumnsWithTypeAndName.h>
@@ -478,12 +480,12 @@ void dumpSessions(const DB::KeeperStorage & storage, const std::string & output_
     }
 }
 
-void dumpNodes(const DB::KeeperStorageImpl & storage, const std::string & output_file, const std::string & output_format, bool parallel_output, bool with_acl)
+void dumpNodes(const DB::KeeperMemoryStorage & storage, const std::string & output_file, const std::string & output_format, bool parallel_output, bool with_acl)
 {
     SharedContextHolder shared_context;
     ContextMutablePtr global_context;
 
-    using PrintFunction = std::function<void(const std::string & key, const DB::KeeperStorageImpl::Node & value)>;
+    using PrintFunction = std::function<void(const std::string & key, const DB::KeeperMemoryStorage::Node & value)>;
 
     const auto print_nodes = [&](const PrintFunction print_function)
     {
@@ -493,7 +495,7 @@ void dumpNodes(const DB::KeeperStorageImpl & storage, const std::string & output
         {
             auto key = keys.front();
             keys.pop();
-            auto value = storage.container.getValue(key);
+            auto value = storage.nodes.container.getValue(key);
             print_function(key, value);
             for (const auto & child : value.getChildren())
             {
@@ -726,7 +728,7 @@ int dumpStateMachine(
     if (dump_sessions)
         dumpSessions(state_machine->getStorageUnsafe(), output_file, output_format, parallel_output);
     else
-        dumpNodes(assert_cast<DB::KeeperStorageImpl &>(state_machine->getStorageUnsafe()), output_file, output_format, parallel_output, with_acl);
+        dumpNodes(assert_cast<DB::KeeperMemoryStorage &>(state_machine->getStorageUnsafe()), output_file, output_format, parallel_output, with_acl);
     return 0;
 }
