+++ src/Coordination/ZooKeeperDataReader.h
@@ -2,6 +2,8 @@
 #include <string>
 #include <memory>
 
+#include <Coordination/KeeperStorage_fwd.h>
+
 namespace Poco
 {
 class Logger;
@@ -12,14 +14,12 @@ using LoggerPtr = std::shared_ptr<Poco::Logger>;
 namespace DB
 {
 
-class KeeperStorageImpl;
+void deserializeKeeperStorageFromSnapshot(KeeperMemoryStorage & storage, const std::string & snapshot_path, LoggerPtr log);
 
-void deserializeKeeperStorageFromSnapshot(KeeperStorageImpl & storage, const std::string & snapshot_path, LoggerPtr log);
+void deserializeKeeperStorageFromSnapshotsDir(KeeperMemoryStorage & storage, const std::string & path, LoggerPtr log);
 
-void deserializeKeeperStorageFromSnapshotsDir(KeeperStorageImpl & storage, const std::string & path, LoggerPtr log);
+void deserializeLogAndApplyToStorage(KeeperMemoryStorage & storage, const std::string & log_path, LoggerPtr log);
 
-void deserializeLogAndApplyToStorage(KeeperStorageImpl & storage, const std::string & log_path, LoggerPtr log);
-
-void deserializeLogsAndApplyToStorage(KeeperStorageImpl & storage, const std::string & path, LoggerPtr log);
+void deserializeLogsAndApplyToStorage(KeeperMemoryStorage & storage, const std::string & path, LoggerPtr log);
 
 }
